// ui3/src/generated/bridge/TravelOutcome.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { Operation } from "./Operation";
import type { TravelPhase } from "./TravelPhase";

export type TravelOutcome = { operation: Operation, phase: TravelPhase, reason: string | null, };